server/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==1.2.0
PyJWT[crypto]==2.8.0
langchain==0.1.12
langchain-community>=0.0.28
cohere==4.52
pydantic==2.5.3
cryptography==44.0.2 